test: extract CIMD suite and shared scaffolding from the provider monolith
oauth-provider.test.ts was at 15,013 lines. The CIMD block was fully
self-contained, so it moves to oauth-cimd.test.ts with the reusable
mocks (MockKV, MockExecutionContext, TestApiHandler, request/env
factories) in test-helpers.ts. Test count is unchanged.
